Update step 3
As with the two previous, this contains similar changes. I also
added a longer explanation for why `bash` is added sometimes, as
well as a small explanation for what the flag -it allows us to do.
Nour also wrote some questions formed as TODO:s, I added my input
to them as well. They can probably be removed, if you dont have
any objections to that!